Synchronize the IrSO functional job with IrSO
* Use minikube instead of kind for the cluster
* Add a JUnit report and annotate failures in the job

+++ hack/ci-e2e.sh
@@ -5,6 +5,8 @@ set -eux -o pipefail
 REPO_ROOT=$(realpath "$(dirname "${BASH_SOURCE[0]}")/..")
 cd "${REPO_ROOT}" || exit 1
 
+CLUSTER_TYPE="${CLUSTER_TYPE:-kind}"
+
 IRSO_REPO="${IRSO_REPO:-https://github.com/metal3-io/ironic-standalone-operator}"
 IRSO_BRANCH="${IRSO_BRANCH:-main}"
 if [[ -z "${IRSO_PATH:-}" ]]; then
@@ -13,12 +15,16 @@ if [[ -z "${IRSO_PATH:-}" ]]; then
 fi
 export IRONIC_CUSTOM_IMAGE=localhost/ironic:test
 
-podman build -t "${IRONIC_CUSTOM_IMAGE}" .
+if [[ "${CLUSTER_TYPE}" == kind ]]; then
+    podman build -t "${IRONIC_CUSTOM_IMAGE}" .
 
-archive="$(mktemp --suffix=.tar)"
-podman save "${IRONIC_CUSTOM_IMAGE}" > "${archive}"
-kind load image-archive -v 2 "${archive}"
-rm -f "${archive}"
+    archive="$(mktemp --suffix=.tar)"
+    podman save "${IRONIC_CUSTOM_IMAGE}" > "${archive}"
+    kind load image-archive -v 2 "${archive}"
+    rm -f "${archive}"
+else
+    minikube image build -t "${IRONIC_CUSTOM_IMAGE}" .
+fi
 
 cd "${IRSO_PATH}/test"
 # shellcheck disable=SC1091
